OH-453: Convert organisaatiotyyppi as koodivalue on conversion services. Move all organisation type conversions from OrganisaatioUtils to OrganisaatioTyyppi.
