+++ organisaatio-service/src/main/java/fi/vm/sade/organisaatio/service/converter/v3/OrganisaatioRDTOV3ToOrganisaatioConverter.java
@@ -15,13 +15,13 @@
  */
 package fi.vm.sade.organisaatio.service.converter.v3;
 
+import fi.vm.sade.organisaatio.api.model.types.OrganisaatioTyyppi;
 import fi.vm.sade.organisaatio.dto.mapping.OrganisaatioNimiModelMapper;
 import fi.vm.sade.organisaatio.dto.v3.OrganisaatioRDTOV3;
 import fi.vm.sade.organisaatio.model.*;
 import fi.vm.sade.organisaatio.resource.dto.OrganisaatioMetaDataRDTO;
 import fi.vm.sade.organisaatio.service.converter.AbstractToDomainConverter;
 import fi.vm.sade.organisaatio.service.util.OrganisaatioNimiUtil;
-import fi.vm.sade.organisaatio.service.util.OrganisaatioUtil;
 import org.apache.solr.common.util.Base64;
 import org.modelmapper.TypeToken;
 import org.slf4j.Logger;
@@ -85,7 +85,7 @@ public class OrganisaatioRDTOV3ToOrganisaatioConverter extends AbstractToDomainC
 
         // t.set(s.getPuhelin());
         s.setToimipisteKoodi(t.getToimipistekoodi());
-        s.setTyypit(OrganisaatioUtil.tyypitToKoodis(t.getTyypit()));
+        s.setTyypit(OrganisaatioTyyppi.tyypitToKoodis(t.getTyypit()));
         // t.set(s.getTyypitAsString());
         s.setVuosiluokat(convertListToList(t.getVuosiluokat()));
         s.setOrganisaatioLisatietotyypit(t.getLisatiedot().stream()
